main.cpp: Ausgaben und Schleife an den Typ vector angepasst.
SearchCART.cpp: Rückgabetyp vector korrekt implementiert (getSize, loader).
loader() auf die Kreise der Hough Berechnung umgebaut. Sofern Kreise gefunden werden, wird der Erste zurück gegeben. Andernfalls der Mittelpunkt des Bildes.

Servo.cpp: Bugfix vertauschte Reihenfolge in initialiseServo().
Servofunktion aktiviert.

+++ src/SearchCART.h
@@ -21,7 +21,7 @@ public:
     SearchCART();
     ~SearchCART();
     vector<int> loader(int i);
-    vector<double> getSize();
+    vector<int> getSize();
     double getFPS();
     double getFC();
     bool isConnected();
